Updated: code to store the shopify identification while creating the order object and declared the method as a helper(#1x634cr)
