+++ src/main.ts
@@ -70,15 +70,6 @@ app.config.globalProperties.$filters = {
       }
     }
     return customerLoyalty;
-  },
-  getIdentification(identifications: any, id: string) {
-    let externalId = ''
-    if (identifications) {
-      const externalIdentification = identifications.find((identification: any) => identification.startsWith(id))
-      const externalIdentificationSplit = externalIdentification ? externalIdentification.split('/') : [];
-      externalId = externalIdentificationSplit[1] ? externalIdentificationSplit[1] : '';
-    }
-    return externalId;
   }
 }
