+++ src/types/Order.ts
@@ -8,7 +8,7 @@ export interface Order {
     itemGroup?: Array<OrderItemGroup>;
     total?: number;
     statusId?: String;
-    identifications?: Array<any>;
+    identifications?: any;
     notes?: Array<any>;
 }
 export interface OrderItem {
